Changed framebuffer, from float (0.0 - 1.0) representation to uint8_t (0 - 255) representation.

+++ tracy.h
@@ -62,7 +62,7 @@ typedef struct Scene3D {
 } Scene3D;
 
 typedef struct Render3D {
-    float* buffer;
+    uint8_t* buffer;
     uint32_t width;
     uint32_t height;
     uint32_t spp;
